# ♻️ Refactor: Improve performance analyseConstantPart