--- a/path.go
+++ b/path.go
@@ -119,12 +119,8 @@ var (
 	routeDelimiter = []byte{slashDelimiter, '-', '.'}
 	// list of greedy parameters
 	greedyParameters = []byte{wildcardParam, plusParam}
-	// list of chars for the parameter recognizing
-	parameterStartChars = []byte{wildcardParam, plusParam, paramStarterChar}
 	// list of chars of delimiters and the starting parameter name char
 	parameterDelimiterChars = append([]byte{paramStarterChar, escapeChar}, routeDelimiter...)
-	// list of chars to find the end of a parameter
-	parameterEndChars = append([]byte{optionalParam}, parameterDelimiterChars...)
 )
 
 // RoutePatternMatch reports whether path matches the provided Fiber route pattern.
@@ -280,22 +276,27 @@ func addParameterMetaInfo(segs []*routeSegment) []*routeSegment {
 // findNextParamPosition search for the next possible parameter start position
 func findNextParamPosition(pattern string) int {
 	// Find the first parameter position
-	nextParamPosition := findNextNonEscapedCharsetPosition(pattern, parameterStartChars)
-
-	// If pattern contains a parameter and it's not a wildcard
-	if nextParamPosition != -1 && len(pattern) > nextParamPosition && pattern[nextParamPosition] != wildcardParam {
-		// checking the found parameterStartChar is a cluster
-		i := nextParamPosition + 1
-		for i < len(pattern) {
-			if findNextNonEscapedCharsetPosition(pattern[i:i+1], parameterStartChars) != 0 {
-				// It was a single parameter start char or end of cluster
-				break
-			}
-			nextParamPosition++
-			i++
+	var search = [256]bool{
+		wildcardParam:    true,
+		plusParam:        true,
+		paramStarterChar: true,
+	}
+	next := -1
+	for i := range pattern {
+		if search[pattern[i]] && (i == 0 || pattern[i-1] != escapeChar) {
+			next = i
+			break
 		}
 	}
-	return nextParamPosition
+	if next > 0 && pattern[next] != wildcardParam {
+		// checking the found parameterStartChar is a cluster
+		for i := next + 1; i < len(pattern); i++ {
+			if !search[pattern[i]] {
+				return i - 1
+			}
+		}
+	}
+	return next
 }
 
 // analyseConstantPart find the end of the constant part and create the route segment
@@ -318,29 +319,46 @@ func (parser *routeParser) analyseParameterPart(pattern string, customConstraint
 	isWildCard := pattern[0] == wildcardParam
 	isPlusParam := pattern[0] == plusParam
 
-	var parameterEndPosition int
-	if strings.ContainsRune(pattern, rune(paramConstraintStart)) && strings.ContainsRune(pattern, rune(paramConstraintEnd)) {
-		parameterEndPosition = findNextCharsetPositionConstraint(pattern[1:], parameterEndChars)
-	} else {
-		parameterEndPosition = findNextNonEscapedCharsetPosition(pattern[1:], parameterEndChars)
-	}
-
+	parameterEndPosition := 0
 	parameterConstraintStart := -1
 	parameterConstraintEnd := -1
-	// handle wildcard end
-	switch {
-	case isWildCard, isPlusParam:
-		parameterEndPosition = 0
-	case parameterEndPosition == -1:
-		parameterEndPosition = len(pattern) - 1
-	case bytes.IndexByte(parameterDelimiterChars, pattern[parameterEndPosition+1]) == -1:
-		parameterEndPosition++
-	}
 
-	// find constraint part if exists in the parameter part and remove it
-	if parameterEndPosition > 0 {
-		parameterConstraintStart = findNextNonEscapedCharPosition(pattern[:parameterEndPosition], paramConstraintStart)
-		parameterConstraintEnd = strings.LastIndexByte(pattern[:parameterEndPosition+1], paramConstraintEnd)
+	// handle wildcard end
+	if !isWildCard && !isPlusParam {
+		parameterEndChars := [256]bool{
+			optionalParam:    true,
+			paramStarterChar: true,
+			escapeChar:       true,
+			slashDelimiter:   true,
+			'-':              true,
+			'.':              true,
+		}
+		parameterEndPosition = -1
+		search := pattern[1:]
+		for i := range search {
+			if parameterConstraintStart == -1 && search[i] == paramConstraintStart && (i == 0 || search[i-1] != escapeChar) {
+				parameterConstraintStart = i + 1
+				continue
+			}
+			if parameterConstraintEnd == -1 && search[i] == paramConstraintEnd && (i == 0 || search[i-1] != escapeChar) {
+				parameterConstraintEnd = i + 1
+				continue
+			}
+			if parameterEndChars[search[i]] {
+				// if (parameterConstraintStart == -1 && parameterConstraintEnd == -1) ||
+				// 	  (parameterConstraintStart != -1 && parameterConstraintEnd != -1)
+				if parameterConstraintStart^parameterConstraintEnd >= 0 {
+					parameterEndPosition = i
+					break
+				}
+			}
+		}
+		switch {
+		case parameterEndPosition == -1:
+			parameterEndPosition = len(pattern) - 1
+		case bytes.IndexByte(parameterDelimiterChars, pattern[parameterEndPosition+1]) == -1:
+			parameterEndPosition++
+		}
 	}
 
 	// cut params part
@@ -420,57 +438,6 @@ func (parser *routeParser) analyseParameterPart(pattern string, customConstraint
 	}
 
 	return n, segment
-}
-
-// findNextCharsetPosition search the next char position from the charset
-func findNextCharsetPosition(search string, charset []byte) int {
-	nextPosition := -1
-	for _, char := range charset {
-		if pos := strings.IndexByte(search, char); pos != -1 && (pos < nextPosition || nextPosition == -1) {
-			nextPosition = pos
-		}
-	}
-
-	return nextPosition
-}
-
-// findNextCharsetPositionConstraint searches the next char position from the charset
-// unlike findNextCharsetPosition, it takes care of constraint start-end chars to parse route pattern
-func findNextCharsetPositionConstraint(search string, charset []byte) int {
-	constraintStart := findNextNonEscapedCharPosition(search, paramConstraintStart)
-	constraintEnd := findNextNonEscapedCharPosition(search, paramConstraintEnd)
-	nextPosition := -1
-
-	for _, char := range charset {
-		pos := strings.IndexByte(search, char)
-
-		if pos != -1 && (pos < nextPosition || nextPosition == -1) {
-			if (pos > constraintStart && pos > constraintEnd) || (pos < constraintStart && pos < constraintEnd) {
-				nextPosition = pos
-			}
-		}
-	}
-
-	return nextPosition
-}
-
-// findNextNonEscapedCharsetPosition searches the next char position from the charset and skips the escaped characters
-func findNextNonEscapedCharsetPosition(search string, charset []byte) int {
-	pos := findNextCharsetPosition(search, charset)
-	for pos > 0 && search[pos-1] == escapeChar {
-		if len(search) == pos+1 {
-			// escaped character is at the end
-			return -1
-		}
-		nextPossiblePos := findNextCharsetPosition(search[pos+1:], charset)
-		if nextPossiblePos == -1 {
-			return -1
-		}
-		// the previous character is taken into consideration
-		pos = nextPossiblePos + pos + 1
-	}
-
-	return pos
 }
 
 // findNextNonEscapedCharPosition searches the next char position and skips the escaped characters